Reorganized helper func locations
